+++ package.json
@@ -2,11 +2,15 @@
   "name": "use-fetching",
   "version": "1.0.9",
   "description": "A react fetch hook to work functional way, even if you have a context.",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tawsbob/use-fetch.git"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prepublish": "tsc"
   },
   "keywords": [
     "react",
